+++ app/ExternalFiles/File_Types/Video.php
@@ -76,26 +76,6 @@ class Video extends File_Types_Base {
 	}
 
 	/**
-	 * Set meta-data for the file by given file data.
-	 *
-	 * @param array $file_data The file data.
-	 *
-	 * @return void
-	 */
-	public function set_metadata( array $file_data ): void {
-		// get the file object.
-		$external_file_obj = $this->get_file();
-
-		// collect meta data.
-		$video_meta = array(
-			'filesize' => $file_data['filesize'],
-		);
-
-		// save the resulting image-data.
-		wp_update_attachment_metadata( $external_file_obj->get_id(), $video_meta );
-	}
-
-	/**
 	 * Return whether this file should be proxied.
 	 *
 	 * @return bool