+++ app/ExternalFiles/Protocol_Base.php
@@ -53,13 +53,6 @@ class Protocol_Base {
 	private bool $queue_mode = false;
 
 	/**
-	 * Marker to force local check of file.
-	 *
-	 * @var bool
-	 */
-	private bool $force_local_check = false;
-
-	/**
 	 * Constructor, not used as this a Singleton object.
 	 *
 	 * @param string $url The URL to use.
@@ -75,24 +68,6 @@ class Protocol_Base {
 	 */
 	public function get_url(): string {
 		return $this->url;
-	}
-
-	/**
-	 * Set to force local check of this file.
-	 *
-	 * @return void
-	 */
-    public function force_local_check(): void {
-		$this->force_local_check = true;
-    }
-
-	/**
-	 * Return whether local check of the file is forced.
-	 *
-	 * @return bool
-	 */
-	protected function is_local_check_forced(): bool {
-		return $this->force_local_check;
 	}
 
     /**
@@ -346,5 +321,49 @@ class Protocol_Base {
 	 */
 	public function set_queue_mode( bool $add_to_queue ): void {
 		$this->queue_mode = $add_to_queue;
+	}
+
+	/**
+	 * Get temp file from given URL.
+	 *
+	 * @param string $url The given URL.
+	 *
+	 * @return bool|string
+	 */
+	public function get_temp_file( string $url ): false|string {
+		// bail if url is empty.
+		if( empty( $url ) ) {
+			return false;
+		}
+
+		// return false in any other case.
+		return false;
+	}
+
+	/**
+	 * Cleanup temporary files.
+	 *
+	 * @param string $file The path to the file.
+	 *
+	 * @return void
+	 */
+	public function cleanup_temp_file( string $file ): void {
+		// bail if string is empty.
+		if ( empty( $file ) ) {
+			return;
+		}
+
+		// bail if file does not exist.
+		if ( ! file_exists( $file ) ) {
+			return;
+		}
+
+		// get WP Filesystem-handler.
+		require_once ABSPATH . '/wp-admin/includes/file.php';
+		\WP_Filesystem();
+		global $wp_filesystem;
+
+		// delete the temporary file.
+		$wp_filesystem->delete( $file );
 	}
 }