+++ app/ExternalFiles/Protocols/Ftp.php
@@ -315,23 +315,12 @@ class Ftp extends Protocol_Base {
 			return array();
 		}
 
-		// get WP Filesystem-handler.
-		require_once ABSPATH . '/wp-admin/includes/file.php';
-		WP_Filesystem();
-		global $wp_filesystem;
-
-		// save this file in a temporary directory.
-		$temp_file = wp_tempnam( $results['title'] );
-		if ( $wp_filesystem->put_contents( $temp_file, $file_content ) ) {
-			$results['tmp-file'] = $temp_file;
-		}
-
 		// get the mime types.
 		$mime_type            = wp_check_filetype( $results['title'] );
 		$results['mime-type'] = $mime_type['type'];
 
 		// get the size.
-		$results['filesize'] = wp_filesize( $temp_file );
+		$results['filesize'] = $ftp_connection->size( $file_path );
 
 		// get the last modified date.
 		$results['last-modified'] = $ftp_connection->mtime( $file_path );